add folder uniform
